Refined version of what is now build_grimoirelab.
The old build_pypi.py is now build_grimoirelab. The program was
restructured and completed a bit. Now works with repositories
that produce several Python packages, or that produce packages
of names different to the name of the repository.
A README.md was also produced, explaining (a bit) the tool.