Add atonce flag.
It seems that when pip installing a set of packages at once,
dependencies among them are not checked. Therefore, we
need a way to pip install one by one, so that they are checked.
This is now the default behavior. If you want to install
all at once, use the --atonce flag.