[tests] Slightly simplified docker-compose.yml
The default configuration for docker-compose exposed the
Kibana and ElasticSearch ports to the host. This causes
trouble if there are local instances of any of them
already running in the host, or if those ports are busy
for some other reason.

To avoid this problem, docker-compose.yml now avoids
to expose those ports. But the functionality provided
by the previous docker-compose.yml file is still
provided by using -f option to docker-compose,
and docker-compose-local.yml, see current version of
README.md for details.