[tests] Fix docker cointainers configuration info
Dependencies could not be installed from wheels because
the wheel package was not installed. This patch installs it.
A new Python dependency is not available in binary wheels, so
it needs to be compiled. But there was no compiler in the
base container. So, this patch installs it.
There are some other (hopefully minor) issues fixed too.