[docker] Add machinery to produce new docker images
Read README.md in this commit for details.

The Dockerfile-factory produces distributable pypi packages.
The packages for elasticgirl.22 were produced using it.
Now, it does some minimum checking, can be extended to improve
that part significantly.

The Dockerfile-installed produces a container with installed
GrimoireLab packages. When run, produces a minimum dashboard
for some GrimoireLab repos, as a proof of concept. Still needs
some work, but it is already useful. I can also be extended
to do some testing, or some other stuff.