[docker] Fix some details of the images.
Now images for grimoirelab/factory, grimoirelab/installed,
and grimoirelab/full are produced from the root dir of the repo.
This allows to have access to the utils and releases directories
when needed.
.dockerignore to avoid uploading to the docker daemon a lot
of stuff in those directories, which is not needed.
Fix to ensure that grimoirelab/full works, by waiting until
the Kibana index in Elasticserarch is ready.
Updated docs.