[docker] Add configuration for new secured container
The grimoirelab/secured Docker container image is a secured version
of grimoirelab/full. It uses the SearchGuard plugin for Elasticsearch
and Kibana to have authentication and permissions.

This patch include also some resctruturing of the files used
by the Dockerfiles, and of the Mordred configuration files,
which are now structured in several pieces.

The list of repositories, identities, etc. are now for CHAOSS,
instead than for GrimoireLab.