[build_grimoirelab] Add check that tools can be run
This is a first try at improving the testing when producing
packages for Pypi. Two new checks have been included:

* That each module can be installed alone (with its dependencies)
in a virtual environment.
* That one executable can be run for each package. Usually,
for each package, the main executable is run with the "--help"
argument.

The structure of the script is now more complete, paving the
way for running the unit tests of each module.