[build_grimoirelab] Add argument for specifying repos to build
In some cases, some of the standard upstream repos,
which are those that build_grimoirelab uses by default,
need to be replaced (or augmented) with other repositories,
such as local clones of the repos, with local modifications
to test.

This patch adds a new argument, reposfile, to specify a JSON
file with a dictionary, in the same format as the variable
for default repos (variable all_repos in build_grimoirelab).