[build_grimorielab] Compute absolute path for dist directory.
In some cases, you may want to pass a relative link for the
dist directory (where Python packages will be created / used).
Unfortunately, when build_grimoirelab uses this directory
while installing in a virtual environment, the path is
different, and relative paths won't work. Therefore, we
need to make the path absolute before it is uses. That's
what this patch does.