Create structure for a grimoirelab Python package.
This package will be mainly for a simple installation
of a GrimoireLab release. The idea is that this package
pulls, via dependencies, from all GrimoireLab packages
in a release, as specified in the requirements.txt file.

Usually, requirements.txt and grimoirelab/_version.py
will be updated in the same commit, thus freezing
a certain version of the packages. Usually, this will
be done after, or in the same commit, the file in releases
corresponding to the release is committed.

Temporarily, while the module name grimoirelab is freed
by the grimoirelab-toolkit package, we're using
grimoirelab_pkg as the name for this module. This will
change to grimoirelab as soon as possible.