Prepare some machinery for including grimoirelab itself in the release
Now, GrimoireLab is a Python package, so that it can be included
in the GrimoireLab release. This package will include as dependenncies
the versions of all other GrimoireLab packages in the release.
The package is built from the grimoirelab directory.

We have a new option in build_grimoirelab to compute the dependencies:
--dependencies, along with another one, --depfile to specify the
file to write them. This allows to compute the specific dependencies
for a certain GrimoireLab release, so that the grimoirelab package
depends on them. This way, installing a certain grimoirelab package
always pulls the GrimoireLab release packages.

This commit also includes a requirements.txt file, built for the latest official
release. This should be a part of the commit changing the release file,
but since this is the first time...

A new file BUILDING.md is also included, with detailed instructions
about how to build a release.

The options for testing in build_grimoirelab now can also use a configuration
directory for the tests. A sample of it is included as docker/testconf.

Some files are changed to adapt to these new functionalities: ansible_release.yml.

The configuration files for analyzing CHAOSS, which are used to produce the
demo dashboard by the grimoirelab/full Docker image, have been updated
(identities.yaml, menu.yaml, mordred-*.cfg, projects.json...) Now, the latest
studies (onion, areas of code, etc.) are also produced for that dashboard.