[third-party] Add nlp-rest-api tool
This code includes the nlp-rest-api tool (available
at https://github.com/crossminer/nlp-rest-api), which
is able to derive sentiment and emotion data from text.

The dockerfile grimoirelab-3p file has been modified to
download and compile this tool. The tool is exposed on
the port 8080 by default, however such a port can be
easily changed when executing the docker run command.

Furthermore, a new entrypoint script is provided. It
extends the entrypoint-full (available in the docker folder)
with the commands needed to start and initialize the
nlp-rest-api service.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>