[docker-compose] Add Filebeat container to ship logs from Mordred to ES
Filebeat container was added to the compose file to ship the logs written by
Mordred to the Elasticsearch cluster. A pretty simple configuration file was added
to the "default-grimoirelab-settings" folder. As a result Filebeat creates
daily indexes in the cluster with the content of the Mordred log file.

The README file was also modified to help the user know how to add the index
pattern needed to visualize the logs.

Signed-off-by: Luis Cañas-Díaz <lcanas@bitergia.com>