[gitmodules] Add GrimoireLab src as git submodules
This commit includes all the source code of GrimoireLab
in a new directory called src.
All the repositories are git submodules and are not
downloaded by default.
To download the contents you need to run `git submodule init`
to initialize your local configuration file, and `git submodule
update` to fetch all the data from the projects.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>