[release] Fix workflow triggering from another workflow
This commit fixes an issue that didn't allow triggering
the `release.yml` workflow from `grimoirelab-release.yml`
when creating a new tag from the same repository.

`actions/checkout` allows to include a token variable that
by default is `github.token` and is persisted in the local
git config, which makes that defining the token in any other
way doesn't use it.

The default `github.token` prevents you from accidentally
creating workflow runs, that's why it didn't trigger a new
workflow when it creates a new tag.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>