[release] Update the way GrimoireLab upgrades its version
This commit updates the way the release workflow checks the
new version number for GrimoireLab.

In the previous version, each package indicated the version
changed, but that method isn't idempotent when the workflow is
restarted.

Now is in the last step when the version is calculated
based on the packages installed in GrimoireLab and the new
version for each package.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>