Fix missing release notes in NEWS file
This commit fixes a bug in the workflow that caused the NEWS file
not to be updated. The issue occurred because a boolean input in
the `if` condition must be compared with a boolean value. Additionally,
when using the boolean input in the run section, it must be compared
as a string.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>