[tests] Add end-to-end performance tests
This is an initial attempt of creating end to end tests.
For this particular case, we focus on performance tests
analyzing git repositories.

The tests create a pre-defined number of repositories
with a random number of commits. The analysis should
be completed before a certain timeout. After the timeout,
the number of events obtained for every repository
have to match with the number of commits in that repository.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>