Update docker configuration to deploy GrimoireLab
Refactor Docker and docker compose configuration to use
the latest version of the containers and use the latest
version of GrimoireLab core. Remove SortingHat as a service
as it is partially included in GrimoireLab core.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>
[ui] Fix taskId type from Int to ID

Fix a bug that caused updating and deleting tasks
from the UI to fail. The GraphQL Task type now uses
an ID field that accepts both integers and strings,
but the UI had not been updated to handle this change.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>